Bug arreglado Int --> float

+++ src/apollo/schema/tpv/tpvDefs.js
@@ -9,7 +9,7 @@ const tpvDefs = (0, apollo_server_express_1.gql) `
         nombre: String
         enUsoPor: Empleado
         libre: Boolean
-        cajaInicial: Int
+        cajaInicial: Float
         createdAt: String
         updatedAt: String
     }
