+++ src/apollo/schema/tpv/tpvDefs.ts
@@ -8,7 +8,7 @@ const tpvDefs = gql`
         nombre: String
         enUsoPor: Empleado
         libre: Boolean
-        cajaInicial: Int
+        cajaInicial: Float
         createdAt: String
         updatedAt: String
     }
